Move common functions to abstract class, make Multipart uploader the same as Uploader
